bdevperf: Change unmap support and I/O size checks to do once per bdev
It is enough to check if unmap is supported and check if I/O size is
multiple of data block size once per bdev.
Hence move these checks from bdevperf_construct_target() to
_bdevperf_construct_targets().

As a result of this change, local variables block_size and data_block_size
are not needed anymore and so use the corresponding APIs directly.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I2cb94c8aa2b6b1a1b9e9a04da533e1fd093f8f29