bdevperf: Use spdk_for_each_channel() for bdevperf_free_targets()
Free each target on the thread which created it by using
spdk_for_each_channel(). This will make possible for us to use
spdk_for_each_channel() for performance dump even for shutdown case
because spdk_for_each_channel() will be serialized if it is called
on the same thread.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Id81a37cd266099206fd38c03aea3a7cfc0a32eb5