Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> So after this patch, do you think we will turn off this AUTOBUILD flag on some of the test systems?
> 
> I just want to make sure we don't end up losing scan-build coverage on any SPDK components.  For example, if we don't do scan-build on any job where SPDK_TEST_RBD is enabled, then we won't get scan-build coverage for the rbd module.

I was planning on turning it off on everything except the unittest job. That is currently the only job on which we run scan-build.
The intention originally was that the unittest job would be run with all flags enabled that triggered a separate compilation, then we would just skip the functional tests so we would still get full coverage on scan-build.
However, that is no longer the case. At least with RBD, it cannot be compiled at the same time as compressdev because of some isa-l conflicts. (I wonder if that's fixable).
Thanks for your incisive comment. I am going to do an audit of the unittest job to make sure it is actually testing everything. If RBD can't be combined with compress I will also add scan-build to the RBD job.
Tomek's comment about printing out information about what we build in scan-build and what we don't is really interesting. I think just like the work I am doing with the regular tests, we should have some kind of automated validation assurance for scan-build coverage. That's the only way we will prevent ourselves from losing coverage in the future.
I also don't think enabling scan-build on every job is a very tenable solution since it takes quite a bit of time, especially on VMs, to run.

Patch-set: 10