configure: add note about changing arch
A reminder to make the same change to DPDK for now. Making this
auto-magic wo;uld be nice but it doesn't follow the same config
type changes we make for other DPDK options and the arch names
they use don't match standard GNU names so we'd need to keep a
table or something and maintain it.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: I9b5adc421e7d732af98a9b4aace68d3e6ec9354d