Update patch set 3
Patch Set 3:

I'm thinking we don't need this patch anymore.  I originally thought this was the patch that we needed, but as you pointed out, the problem was really more with the req allocation, not the submission.  We can safely rely on the normal request queueing if we get an EAGAIN on either of the fused commands.

What do you think?

Patch-set: 3
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>