Update patch set 1
Patch Set 1:

Crap - I just thought of another problem.  This doesn't work if the request needs to be split for some reason - in that case, we need more than just one request object.

Let's back these nvme request patches out for now.  I'd like to get the rest of this series moved forward.  I need to think about how to handle this.  For now let's just assert(false) if comparev succeeds and writev fails.

One idea I had was that if the spdk_nvme_ns_cmd_comparev succeeds, but the writev fails with -ENOMEM, that we return -ENOMEM to the bdev layer but save some kind of status in the driver_ctx to say that we already submitted the compare operation.  The problem is how to know whether that status is valid (the driver_ctx could have garbage in it from some other bdev_io operation from some other bdev module).  This is the idea I'm still pursuing though - it would make a lot of this cleaner I think.

Patch-set: 1
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>