bdevperf: Unregister pollers immediatley when target exits with error
When the target ends with error, unregister pollers of the target
immediately by replacing setting target->is_draining to true by
_end_target(target).  Otherwise, when target->run_timer is expired,
end_target() would be called after target is freed and would cause
use-after-free error.  We had seen this failure for a single case
but applies the fix for all cases.

Fixes issue #1140

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: If39779817ed99c9441e368c1d4512c66d13d3378