Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Hi Ben,

 For TCP transport is not same with the RDMA transport. During the negotiation with RDMA transport, the initiator will send the queue size in the beginning. 
But for TCP transport, it will not do this. But it send the sq size during after the TCP connection is constructed. 
In my mind, we should resize the queue size which is required by the initiator. And the current code does not affect performance. And for the SPDK users which want to continue development, they can easy to do debug with the right size in the target. 

Otherwise, you configure a big size, and there will be 
more resources providing to a connection, and many bugs will be concealed. For example, if the initiators requires QD=32, and we always allocate 128 QD. Then there will be more TCP requests objects are created. And it can conceal lots of potential bugs in our code, which can be hard to find especially when we introduce the real async manner. In some cases, we should wait for the TCP req objects, but we do not do it.

Patch-set: 2
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1