Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi Ben,
> 
> Another Question, what's performance benefit while using zero_copy? Do you see some improvements?

At 4k I/O size it is currently no benefit on Xeon servers. But the cost of pinning is mostly a fixed cost and doesn't scale with I/O size like copying does, so at 8k this nearly doubles the performance, and at 16k it is 4x, etc.

But then there's also the new pinning API in io_uring which could potentially eliminate the cost of the pinning. I think at that point, even at 4k we can expect to see big gains.

Patch-set: 4