Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Hi Shuhei,

Is it possible to combine any of the patches in this series?  This patch needs to be updated, which means that rest of the series will also need to be rebased.  A lot of the patches here seem very small - it might be better to combine some of them.  But of course please use your best judgment on which ones should stay separated.

-Jim

Patch-set: 1
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1