lib/blob: add option to disable extent pages
This is an additional option that can be passed when creating
a blob.

Extent pages will contain extents serialized in pages
separate from the md chain of a particular blob.
Opening a blob shall check if SPDK_MD_DESCRIPTOR_TYPE_EXTENT_RLE
was used to persist the clusters. If yes, disable_extent_pages
should be set to true.

Extent Table added in later patches will always be present in md,
even if length is 0 and blob is thick provisioned. This is not true
for Extent RLE, as there can be none in md.
Because of the above on _spdk_blob_alloc() it is always assumed
that disable_extent_pages is set to true.

Later in the series spdk_blob_opts_init() passed when creating
a blob, will set disable_extent_pages to true as default.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I2146da6130a0645e686ab02a3b5d2d86a7d35a1f