lib/thread: Assert if spdk_put_io_channel() is called on the wrong thread
spdk_put_io_channel() has been designed to be called on the same
thread that called spdk_get_io_channel().  However a workaround has
been implemented such that message is sent to the correct thread
if called on the wrong thread.

Bdevperf tool had a design flaw that needed the workaround, but
it was fixed recently.  We do not know if we have any other case.

Hence by adding assert to spdk_put_io_channel(), let's find any
other case and fix it if found.

Besides, correct the comment for spdk_put_io_channel() in
include/spdk/thread.h not to create any other case in future.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I6ec7bf074818abef43b23ca40bc9385adac70a75