Update patch set 5
Patch Set 5:

> Patch Set 4: Code-Review-1
> 
> I'm following why you're using #if 0 here? I assume you're not proposing to merge like this, why not just delete the code if not used here or in a subsequent patch?

Sorry, missed the comment when replaying to previous ones. It was never intention to just straight up remove the original code.
Originally we wanted to keep the code for extents_rle (old ones) for compatibility reasons.
Ben brought up a compelling reason to actually give a choice in that matter to the user.
New extent table/extent pages require at least two pages (md + extent page) even if the blob is quite small. For use cases like rocksdb/blobfs it is very likely for there to exist many of those.

Because of the above, I've now changed this patch to provide creation option to disable the new way to handling metadata. Cases like blobfs will sometimes want to set disable_extent_pages to true.

As a work in progress - this option is set to true right now. When later patches have the path ready, default will be changed to false (and kept that way).

Patch-set: 5