Update patch set 16
Patch Set 16:

> Patch Set 15: Code-Review+1
> 
> hey, that's 502 better than 10 ;)

The low amount of was because it exposed a LOT of issues with serialization/parsing when there were many pages. Which helped when developing 😊

Besides, this change and setting right number of Extents per Extent Page is quite important so that is why it was taken out of the rest of series.

Patch-set: 16
Reviewer: Gerrit User 1011221 <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>