Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> Hi Ziye,
> 
> Can we set pdu->conn in spdk_iscsi_conn_write_pdu(conn, pdu) instead?
> Looks natural to me, and if we do this, we can squash these three patches into a single patch.
> Will you correct me if I'm wrong?
> 
> Thanks,
> Shuhei

Hi Shuhei,

Thanks for your comments. Actually, I compared the code in NVMe TCP. I think that for every pdu after allocated, it should be associated with a connection (There is not meaningful, if a pdu does not belong to any conn). This the reason I add this, since later we may also need to track the used pdu in iSCSI connection like what we do in NVMe-oF. This is why I create a seperate patch.

Patch-set: 9