Update patch set 16
Patch Set 16: Code-Review-1

> Patch Set 15: Code-Review+2
> 
> Hi Ben, 
> Will you check this patch again?
> I checked the specification again and I didn't find obvious issue with this patch.

I've looked this over and I see a few major issues (that I think are also broken in DPDK).

The first has to do with out of order completions. Say two descriptors are submitted to the ring such that the first descriptor goes to the end of the ring and the second one loops back and is placed in the first slot. And then the target picks them up and the backing storage device completes them in reverse order, such that the descriptor in the first slot is completed prior to the descriptor in the last slot, then the wrap counters will be all wrong. The used wrap counter will contain the phase value from the previous iteration through the ring still, and the first descriptor will not appear to be complete, hanging the driver. This is also broken in DPDK's implementation from what I can see, and is actually wrong in the specification too.

The second has to do with the logging of descriptors when a whole chain is completed. Right now it only logs the first descriptor in the chain. I don't know what this logging is for, but my guess is live migration. Does live migration work with packed rings?

Third, our copy of the DPDK code hasn't been updated to include all of the packed ring stuff that DPDK has added (so our internal copy of rte_vhost_set_vring_base, for example, does the wrong thing). There's actually this really weird code set up where there's an SPDK virtqueue object but also a DPDK virtqueue object and they both hold the same values and are used at the same time. That doesn't make any sense to me.

I think adding packed virtqueue support to SPDK first involves writing a unit test to confirm DPDK's implementation is broken in the way I outlined above. If it is, then we first have to fix DPDK's implementation to handle out of order completions. They probably just don't get those in networking. Once those patches are all merged, we'd need to update (or drop entirely) the SPDK internal copy of the DPDK code. Only then can we add actual support for packed virtqueues into SPDK.

Depending on the outcome of the unit test, we'll make a decision on whether to continue to pursue this for 20.01 or even 20.04.

Thanks for all your work here Jin.

Patch-set: 16
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1