test/hello_sock: Reactor the program and add -N option
This option is used to make hello_sock can have use
the designated sock implementations. We need this patch
since we will provide another uring implementation.

So better to pass the name. Otherwise if the users have
many different implemenations, VPP implementation
could not be the highest priority for test.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I79a9e5d7d8e6a90f6baeec7d8bd8f447d3df585a