vagrant/scripts: [RFC] Be AppArmor aware (libvirt)
On systems which are managed by the AppArmor subsystem (e.g. Debian 10)
the libvirtd daemon executes its libvirt-aa-helper to dynamically
create an apparmor profile for each qemu instance. However, currently,
the helper is not taking the <qemu:commandline> into the account and
as a result the nvme disk image is not being included in the said
profile - the emulator, when execed, fails with -EPERM while trying to
access the image.

To mitigate the above, try to detect if nvme image is created under
libvirt's confined environment and if so, patch its main apparmor
profile so it includes the $nvme_disk path.

Change-Id: I674e13952d44fc3773ea1734cebcea37539eca58
Signed-off-by: Michal Berger <michallinuxstuff@gmail.com>