Update patch set 11
Patch Set 11:

> Patch Set 11: Code-Review-1
> 
> This implements write batching, but doesn't use an actual async interface in vpp. The reason batching is important for posix is to reduce syscalls. However, there are no syscalls in vpp, so it's not at all clear to me that this is a performance improvement. Have you benchmarked?
> 
> In general, this async interface should call one of the actual async vpp interfaces directly instead of batching. Until this is thoroughly benchmarked and experimented with, I don't think we should proceed here.

Hi Ben,

This is not for performance purpose at the first, we can continue optimizing this later. I did not test VPP's performance.


The reason is that : If we change writev usage in iSCSI from sync to async, it relies on the VPP support. If no such support, we cannot make iSCSI to use async manner.

Patch-set: 11