Update patch set 4
Patch Set 4:

Hi Vitaliy, Machek,

I'm still work in progress but we will need to have default reactor threads for now.

If we move poll group to dynamic lightweight threads, we have to move login acceptor (NVMf says it listener) to dynamic lightweight threads. This will require us to communicate JSON RPC handler with login acceptor threads. This is not easy task. So for long term, this is what we need, but for short term, this is too hard.

Thanks,
Shuhei

Patch-set: 4
CC: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>