Update patch set 27
Patch Set 27:

(1 comment)

> Patch Set 25: Code-Review-1
> 
> (1 comment)
> 
> @Seth Just to be sure of the comment I made in the patch, I've ran this through nightly job.
> That the job passed where relevant (hit intermittent failure :/).
> https://dqtibwqq6s6ux.cloudfront.net/public_build/autotest-nightly_1411.html
> 
> The output from autorun_post.py show that no tests were executed. Not sure if this is direct result of this patch, previous ones or such behavior occurs when there is failure in test.
> 
> 10:14:01 -----Tests Executed in Build------
> 10:14:01 
> 10:14:01 
> 10:14:01 -----Tests Missing From Build------
> 10:14:01 alias_rpc
> 10:14:01 asan
> 10:14:01 autobuild
> .....
> 
> Can you please check what is the cause and how this patch interacts with jobs outside of per-patch (nightly/vs_dpdk etc..) ?

Thanks Tomek, that's a really great review. It took me a bit, but I finally have a coherent answer for you. The reason that nightly is not reporting any completions on this patch is as follows. Nightly was only configured to run on master, so the way it handles checkout of patches is a little different. It turns out that when you specify a patch other than master, the master node still clones master, but all of the slaves pick up the version of spdk from your patch. This means that all of the slave nodes were creating test_completions files based on this new series, while the master node was using the old version of the script to parse those completions.

I will work with Karol offline to figure out how to fix the nightly configuration, but in the meantime, we shouldn't let that configuration hold up merging these patches as they will work properly once merged to master.

Also, I fixed my patches based on your other comment.

Patch-set: 27