Update patch set 4
Patch Set 4:

We don't need to do this refactor to move ahead on iSCSI writev_async. Can you move this patch instead to the front of the io_uring series and we'll continue to iterate on it there?

Patch-set: 4
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>