Update patch set 27
Patch Set 27: Code-Review+1

This is all OK so +1, but I have one small request to make. ANd I'll even maybe do it myself, but just for the record:

If we ever run a CI instance with a smaller test job subset then we will get fails on each build because not all of the expected tests were run. We should add an option to autorun_post.py which allows to skip verifying which tests were run.

Patch-set: 27
Reviewer: Gerrit User 1011256 <1011256@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1