Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Another concerns from me:
> 
> 1 With this change, for C2hdata sending, we will only support sending one C2hdata PDU (From the spec, if a flag is not set in C2HDATA pdu, the target needs to send C2HDATA pdu + rsp pdu). If the initiator does not support "fast c2hdata recving", it means there will be no rsp pdu generated, then the initiator will hang (Though this should be the initiator fault).

You're right - there is a problem here. If the initiator does not support this optimization, we need to send two PDUs to complete a read. However, since there are no initiators without support for that optimization, we probably don't have to make that code path high performance. I may think about doing the C2H PDU and then the RSP PDU in serial.

> 
> 2 With this change, we will always send one C2hdata pdu. If there is restrict size of PDU size, we cannot send multiple C2hdata pdus. (Currently, there is no restriction of C2hdata restriction size in SPEC). If this is added, we may still need add this back.

This won't be added to the spec - the initiator does not get to flow control the target.

> 
> 3 Our target will support 1 R2T PDU.

Correct

Patch-set: 4